[CSL-1926] Reorganise the modules into Request & Response ones
