+++ wallet-new/server/Cardano/Wallet/API/V1/Handlers/Wallets.hs
@@ -5,7 +5,8 @@ import           Universum
 import qualified Pos.Wallet.Web.ClientTypes.Types as V0
 import qualified Pos.Wallet.Web.Methods as V0
 
-import           Cardano.Wallet.API.Request.Pagination
+import           Cardano.Wallet.API.Request
+import           Cardano.Wallet.API.Response
 import qualified Cardano.Wallet.API.V1.Handlers.Accounts as Accounts
 import           Cardano.Wallet.API.V1.Migration
 import           Cardano.Wallet.API.V1.Types as V1
