+++ wallet-new/src/Cardano/Wallet/API/Response.hs
@@ -1,0 +1,75 @@
+{-# LANGUAGE DeriveGeneric #-}
+module Cardano.Wallet.API.Response (
+    Metadata (..)
+  , OneOf(..)
+  , ExtendedResponse(..)
+  , respondWith
+  ) where
+
+import           Universum
+
+import           Data.Aeson
+import           Data.Aeson.TH
+import qualified Serokell.Aeson.Options as Serokell
+import           Test.QuickCheck
+
+import           Cardano.Wallet.API.Request (RequestParams (..), ResponseFormat (..))
+import           Cardano.Wallet.API.Request.Pagination (PaginationMetadata (..))
+
+-- | Extra information associated with an HTTP response.
+data Metadata = Metadata
+  { metaPagination   :: PaginationMetadata
+    -- ^ Pagination-specific metadata
+  } deriving (Show, Eq, Generic)
+
+deriveJSON Serokell.defaultOptions ''Metadata
+
+instance Arbitrary Metadata where
+  arbitrary = Metadata <$> arbitrary
+
+-- | An `ExtendedResponse` allows the consumer of the API to ask for
+-- more than simply the result of the RESTful endpoint, but also for
+-- extra informations like pagination parameters etc.
+data ExtendedResponse a = ExtendedResponse
+  { extData :: a        -- ^ The wrapped domain object.
+  , extMeta :: Metadata -- ^ Extra metadata to be returned.
+  } deriving (Show, Eq, Generic)
+
+deriveJSON Serokell.defaultOptions ''ExtendedResponse
+
+instance Arbitrary a => Arbitrary (ExtendedResponse a) where
+  arbitrary = ExtendedResponse <$> arbitrary <*> arbitrary
+
+-- | `PaginationParams` is datatype which combines request params related
+-- to pagination together
+
+
+-- | Type introduced to mimick Swagger 3.0 'oneOf' keyword. It's used to model responses whose body can change
+-- depending from some query or header parameters. In this context, this represents an HTTP Response which can
+-- return the wrapped object OR the ExtendedResponse.
+newtype OneOf a b = OneOf { oneOf :: Either a b } deriving (Show, Eq, Generic)
+
+instance (ToJSON a, ToJSON b) => ToJSON (OneOf a b) where
+  toJSON (OneOf (Left x))  = toJSON x -- Simply "unwrap" the type.
+  toJSON (OneOf (Right x)) = toJSON x -- Simply "unwrap" the type.
+
+instance (Arbitrary a, Arbitrary b) => Arbitrary (OneOf a b) where
+  arbitrary = OneOf <$> oneof [ fmap Left  (arbitrary :: Gen a)
+                              , fmap Right (arbitrary :: Gen b)]
+
+
+respondWith :: (MonadPlus p, Monad m) => RequestParams -> m (p a) -> m (OneOf (p a) (ExtendedResponse (p a)))
+respondWith RequestParams{..} generator = do
+  theData <- generator
+  case rpResponseFormat of
+    Extended -> return $ OneOf $ Right $
+      ExtendedResponse {
+        extData = theData
+      , extMeta = Metadata $ PaginationMetadata {
+          metaTotalPages = 1
+        , metaPage = 1
+        , metaPerPage = 20
+        , metaTotalEntries = 3
+      }
+      }
+    _ -> return $ OneOf $ Left theData
