+++ wallet-new/src/Cardano/Wallet/API/V1/Accounts.hs
@@ -4,6 +4,7 @@ import           Universum
 
 import           Servant
 
+import           Cardano.Wallet.API.Response
 import           Cardano.Wallet.API.V1.Parameters
 import           Cardano.Wallet.API.V1.Types
