+++ wallet-new/src/Cardano/Wallet/API/V1/Addresses.hs
@@ -3,6 +3,7 @@ module Cardano.Wallet.API.V1.Addresses where
 
 import           Servant
 
+import           Cardano.Wallet.API.Response
 import           Cardano.Wallet.API.V1.Parameters
 import           Cardano.Wallet.API.V1.Types
